backend/requirements.txt
Flask
flask-cors
google-generativeai
python-dotenv
BioPython
pinecone-client
sentence-transformers
